rduxgetAll

+++ controllers/notes.js
@@ -3,6 +3,9 @@ const Notes = require("../models/Notes");
 exports.getAll = async (req, res) => {
   try {
     const notes = await Notes.find({ user: req.user._id });
+    if (!notes) {
+      res.status(404).send({ msg: "Note not found" });
+    }
     res.status(200).send({ msg: "All your notes", notes });
   } catch (error) {
     res.status(400).send({ msg: "Fail to get " });
@@ -25,10 +28,14 @@ module.exports.addNote = async (req, res) => {
 exports.getNote = async (req, res) => {
   try {
     const note = await Notes.findOne({ _id: req.params.id });
-
-    res.status(200).send({ msg: "The note is:", note });
-  } catch (error) {}
-  res.status(400).send({ msg: "Fail" });
+    if (!note) {
+      res.status(404).send({ msg: "Note not found" });
+    } else {
+      res.status(200).send({ msg: "The note is:", note });
+    }
+  } catch (error) {
+    res.status(400).send({ msg: "Fail" });
+  }
 };
 
 exports.updateNote = async (req, res) => {
@@ -36,13 +43,15 @@ exports.updateNote = async (req, res) => {
     const { title, content, category } = req.body;
     const note = await Notes.findById(req.params.id);
     if (note.user.toString() !== req.user._id.toString()) {
-      res.status(401).send("You can't do this action");
+      return res.status(401).send("You can't do this action");
     }
-    note.title = title;
-    note.content = content;
-    note.category = category;
-    await note.save();
-    res.status(200).send({ msg: "Note Updated successfully...", note });
+    if (note) {
+      note.title = title;
+      note.content = content;
+      note.category = category;
+      await note.save();
+      res.status(200).send({ msg: "Note Updated successfully...", note });
+    }
   } catch (error) {
     res.status(400).send({ msg: "Not Updated!!!" });
   }
@@ -53,13 +62,14 @@ exports.deleteNote = async (req, res) => {
     const note = await Notes.findById(req.params.id);
 
     if (note.user.toString() !== req.user._id.toString()) {
-      res.status(401).send("You can't do this action");
+      return res.status(401).send("You can't do this action");
     }
+    if (note) {
+      const { _id } = req.params;
+      await Notes.deleteOne({ _id: req.params.id });
 
-    const { _id } = req.params;
-    await Notes.deleteOne({ _id: req.params.id });
-
-    res.status(200).send({ msg: "Note deleted successfully..." });
+      res.status(200).send({ msg: "Note deleted successfully..." });
+    }
   } catch (error) {
     res.status(400).send({ msg: "Not deleted!!!" });
   }
